Proxy program -> thread.
